update type to integer

+++ src/hub/dataindex/indexer.py
@@ -5,7 +5,7 @@ class TaxonomyIndexer(indexer.Indexer):
     def __init__(self, build_doc, indexer_env, index_name):
         super().__init__(build_doc, indexer_env, index_name)
         self.es_index_mappings['properties'].update({
-            'lineage': {'type': 'long'},
-            'children': {'type': 'long'},
-            'ancestors': {'type': 'long'}
+            'lineage': {'type': 'integer'},
+            'children': {'type': 'integer'},
+            'ancestors': {'type': 'integer'}
         })
